Revert "Revert "Refactor main.py""
This reverts commit 3537a15754ac2528d3579d89c9a969f2e5645100.
