+++ requirements.txt
@@ -3,3 +3,5 @@ salsa20~=0.3.0
 tabulate~=0.8.10
 pandas~=1.5.3
 scipy~=1.9.0
+
+tornado~=6.3.2
